Make counter read functions visible
In my work on erl-optics, I found that I would like to have the ability to read
the values of counters for testing purposes. This was previously not possible
as the functions were not declared in optics.h.

This commit moves the declarations of the appropriate counter reading functions,
as well as optics_epoch, into optics.h so that they may be used in libraries
that link against liboptics.